RAG_try/requirements.txt
flask==2.3.3
werkzeug==2.3.7
langchain>=0.1.0
langchain-core>=0.1.52
langchain-community>=0.0.13
langchain-huggingface>=0.0.2
openai==1.3.7
transformers==4.36.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
torch==2.2.0
numpy>=1.26.0
pypdf==3.17.1
huggingface-hub>=0.23.0
tensorflow==2.18.1 